requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langchain-groq==0.1.5
langchainhub==0.1.15
notebook==7.1.3
pinecone-client==3.2.2
pypdf==4.2.0
chainlit==1.1.306
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.66.4
requests==2.31.
fastapi==0.112.0